Refactored code for more clarity
